split file into modules and combine all typing declaration
