+++ src/options.ts
@@ -20,3 +20,26 @@ export interface FieldOption {
   aggregate?: string;
   bin?: boolean;
 }
+
+export interface SupplementedFieldOption extends FieldOption {
+  removeOriginalTemporalField?: string;
+  bin?: boolean;
+}
+
+export type TooltipData = {title: string, value: string | number };
+
+export type Scenegraph = {
+  datum: {
+    _facetID: number,
+    _id: number
+  },
+  mark: {
+    marktype: string,
+    items: object,
+    name: string
+  }
+};
+
+export type VgView = any;
+
+export const DELAY = 100;
