--- src/supplementedFieldOption.ts
@@ -1,5 +1,0 @@
-import {FieldOption} from './options';
-export interface SupplementedFieldOption extends FieldOption {
-  removeOriginalTemporalField?: string;
-  bin?: boolean;
-}
